safety_tool: move utils_cmd to utils module as utils::cmd
