+++ safety-tool/src/utils/cmd.rs
